stop sending data when stream is ended
server may respond to requests before it gets all data, for example by
analyzing headers. In such situation we have pending_body around to send.
We drop it and send end_stream frame that is not subject of flow control.

Without this code we call release callback and parent class thinks we
have stream available and sends next request in, but H2 considers the
stream as half closed and we get TooManyStreamsError

see also https://github.com/python-hyper/hyper-h2/issues/544